# Conflicts:
#	platform/src/main/java/ai/enpasos/muzero/platform/agent/memorize/ReplayBuffer.java
#	platform/src/main/java/ai/enpasos/muzero/platform/run/train/MuZero.java


Former-commit-id: 386082df9f7b4d2da4ac853a76f2a94cc910a066

+++ platform/src/main/java/ai/enpasos/muzero/platform/common/Functions.java
@@ -76,7 +76,7 @@ public class Functions {
     }
 
     public static double entropy(List<Float> ps) {
-        return (double)ps.stream().reduce(0f, (e, p) -> e + (float)singleEntropySummand(p));
+        return ps.stream().reduce(0f, (e, p) -> e + (float)singleEntropySummand(p));
     }
 
     public static double singleEntropySummand(double p) {
